pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eyra"
version = "1.0.0"
description = "Real-time AI screen analysis from the terminal."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Soroush Yousefpour" }
]
keywords = ["ai", "screen", "vision", "macos", "local", "privacy", "ollama", "openai-compatible"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0.0",
    "mss>=10.1.0",
    "opencv-python>=4.11.0",
    "pillow>=11.0.0",
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/gabrimatic/eyra"
Repository = "https://github.com/gabrimatic/eyra"
"Bug Tracker" = "https://github.com/gabrimatic/eyra/issues"
Changelog = "https://github.com/gabrimatic/eyra/blob/master/CHANGELOG.md"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["src/chat", "src/clients", "src/modes", "src/utils", "src/runtime"]

[tool.hatch.metadata]
allow-direct-references = true
